Add configuration package
This project will make use of a configuration package for breaking down
ValueSets, StructureDefinitions, and CodeSystem definitions into
templated solutions for generating FHIR documentation and libraries.

This adds the basic YAML parsing support to the project for this, as
well as an example for what this format will look like.