Add template-wrapping packages
This project makes heavy use of templates for the inputs, and it's
going to be much easier to define a set of template functionalities
that we want now in a single place. To do this, a `template` package
is added that adds standard `text/template`-like wrapping around
templates, with a series of default helper functions installed.