Add workflow check at end of presubmit
GitHub requires explicit named workflow jobs in order to be marked as
"required", however there are a dynamic number of jobs in presubmit
that ultimately need to pass. Rather than forcing "all" workflow jobs
as being marked as required, it's easier to have a single job that
depends on all other jobs as passing in order to be marked as passing.

This adds that no-op job so that we can wait on it on CI.