Add package for loading FHIR IG data
In order to read all structure definitions of an IG, it's important to
be able to actually have a common location for all of this to be.

As it turns out, the simplifier.net API is actually really
straight-forward to fetch content, including all dependencies
recursively. Thus, this adds a small integration for it that includes
a package cache and filtering.

The current testing practices could be better, but it's a start, and
it works.