Update configuration to use one template
The configuration was previously supporting header and footer templates
with the goal of enabling "single" headers/footers for the same file.
However, refactorings have now made the template take in a series of
definitions instead of a single definition. This means that the
configuration can be simplified to use a single template.