Add crude model-graph building implementation
This is a first-draft, which will not be the final implementation. This
exists in its current fashion so that this can be used to get closer
to generated Resource bindings in Go. Once we achieve that, this code
will be replaced with something that uses itself to bootstrap the
parsing and management.

We can live with a little cruft to start this off. An unfortunate
consequence is that, for the time being, there will be less testing
as a result. This is a trade-off that we are willing to make for now,
but is not suited for long-term generation.