Implement second version of modeling
The initial modeling was unwieldy and difficult to work with. It made
testing complicated -- which was effectively completely skipped, and
is not a good practice. This commit introduces a new version of the
modeling that is more flexible and easier to work with -- which will
replace the previous model in the following commits.