Rework generation process to use new model
This is a larger commit that removes all the previous modeling logic
in favor of the new model. In the process, this updates generation to
no longer allow sequences of entries, and to instead enforce only a
single value per generated file -- which is ultimately much easier
to read given the content size.

This decision may be changed again in the future, so this is not a
stable API.