Remove unused model func in template
The template previously was allowing for a series of context-sensitive
model functions to be installed; however this actually poses a
challenge for the "template" API, which requires that functions be
installed _before_ being parsed. Parsing happens upon deserialization,
which causes these functions to not be able to be installed at all.

Ultimately the change to using only a single value for the template
actually simplifies this, since we can now easily capture the context
of the input value to the template -- rather than requiring a new
function that exists solely to return the value.