Allow defining multiple templates per transform
Defining multiple templates makes it easier to reuse common component
across expansion -- such as trying to come up with a "name" for a type
as part of the template expansion. This just drastically simplifies
things and reduces duplication.

Additionally, paths are now forced to be relative to the config file,
which is more intuitive.