Add 'default' config node for shared defaults
The current template mechanism only allows for templates to be specified
per transformation. For shared template segments, this means it gets
specified for each value -- which is actually a lot of times in some
cases.

To make this a little simpler, this adds a 'default' config node that
can be used to specify shared defaults for all templates.