Add `--rm` flag for deleting output dir
Executing `fhenix run` is currently dangerous because it will delete
the contents of the output directory by default without asking. A
consequence of this is that I actually blew away an entire project,
git directory and all, by trying to generate contents to the root.

That's a bad situation to be in, and a bad default to have. To address
this I've added an `--rm` flag that is not default, and must be
specified in order to delete content first.

Fixes: #6