Add build-docs to presubmit
The build-docs stage has been failing due to various minor problems,
largely encountered due to the docs deployment workflow not actually
being tested before promotion. This now pulls out the building of the
docs into a reusable workflow that is embedded in the `presubmit`
pipeline to ensure that the next and future promotions will actually
be able to generate proper documentation.