Add workaround for positiveInt and unsignedInt
The r4 FHIR spec defines the positiveInt and unsignedInt as having a
value type representation of _string_ rather than as integer. This is
incorrect modeling that was fixed in r4b, but not retroactively
addressed.

This fixes the issue by adding an explicit workaround only for the
r4 FHIR IG.