Add internal v1 configuration package
As this project moves closer to a v1, the format for the configuration
has become more stable. This commit adds a new internal package that
will be used to parse and validate the v1 configuration format.

This package is purely internal, and will be used by a public 'config'
package in a following commit.