Add `transform` package
The transformation logic from model format into a template output format
has not been implemented that well. The first implementation was done
largely in-place to make it easy to implement -- but it's not a good
architecture. The templates force a _single_ input value and disallows
anything larger.

This seeks to correct that by converting the new configuration format
into something that is aware of the underlying type(s) in the model,
and can join large quantities of definitions into the same output format.