Add `input` node to base configuration
The `v1` configuration was previously missing the `input` field in
parsing, and missed from being included in the transformed configuration
as well.

In the process, the transformed output has been changed to refer to the
package content rather than the listed files -- this way configuration
handling isn't responsible for fetching file content.