Add `fhirsource` package
This tool wants to support both having local and remote package
definitions that can be used as input into the generation mechanism.
To support that, this introduces a new package `fhirsource` which
operates off of the existing configuration format in order to fetch
all content -- either from a remote or from a local source.