Export `model` package
As the public pieces of this project come together with the new
overhauled configuration packages, we are in need of some public type
for the "model" so that we can easily set up transformation objects.

This promotes the existing internal model to public. This is currently
not a stable API, as it will be changing soon as this gets reworked to
use the generated output content.