Fix panic condition in legacy config
The legacy configuration may panic if no "partial" is specified in the
yaml file. This occurs due to the partials map being `nil` after
yaml decoding -- but the logic attempts to provide defaults if
unspecified, which causes a nil-access violation.

The fix being added her is just to create the map first.