Add definitional conformance model
Eventually the goal is to fully replace the existing `model` package
with something that is easier to use and test. As an incremental step
there, we need a way that can be used to define conformance
types from their JSON definitions.

This introduces the initial version which does precisely that; it
parses the JSON definitions and creates the appropriate resource types
that are part of the conformance model.