Add package for conformance module
FHIR defines a conformance module, which effectively provides the
skeleton of all definitional resources that define the size, shape,
and interaction points with the FHIR standard being used.

This attempts to create a collection of these types in a reasonable
way so that a new model mechanism can be built on top of it.
In particular, this handles the parsing, filtering, and iteration of
the conformance resources.