Rework model to use fhir package
The model was initially bootstrapped with custom JSON definitions of the
Resource structures; however, since `fhenix`'s 0.1 version, the FHIR
types have been generated from the structure-definitions and have allowed
bootstrapping the `fhenix` system to be in terms of this.

This commit reworks the model to use the generated FHIR types from the
`fhir` package, instead of the hand-authored ones.