Fix "too many files" failure
Fhenix has been sporadically failing due to "too many files" being open
when building on some systems in CI. The exact cause has been elusive,
but it looks like it is due to the custom read/write stream type not
being closed -- each instance of which creates 2 new leaking file
descriptors: a read and a write descriptor. This adds a proper `defer`
command to close the file upon completion to fix the issue.