Remove teeReadCloser type
The initial implementation made use of a custom `teeReadCloser` type,
which acted just like the `io.TeeReader`, but included the `Close`
function in the interface. In hindsight, this is actually not a necessary
requirement -- since the package files are already closed outside the
use of this type.

This removes the type.