Add 'registry' package for fhir-ig access
The current registry-handling logic in Fhenix is... well, it could be
better. It was written in a hurry to get a first draft of this project
working -- but wasn't designed to be particularly testable or robust.

This commit tries to change that. It doesn't *yet* replace the internal
fhirig logic -- but it starts it off by providing a full registry client
abstraction with a full test-double package to test against it. This
makes it really easy to test both ends of the integration, since
packages can be articulated in either the client or the fake server.

This will be a pretty large change overall, so more parts of this will
be coming in later commits to provide a full replacement for the existing
fhirig logic.