Fix 'too many open files' for the last time
The 'too many open files' issue has come back again. It seems my initial
code for this was way too sloppy and left a lot of file descriptors open
when it shouldn't be -- and this is breaking on macOS.

It turns out, the output from the `os.Create` command during unpacking
was missing being closed -- and this has been contributing to the issue.
Hopefully this is the last missed descriptor.