Add `registry.Package` type
The previous changes have been incrementally building up a registry
abstraction for managing packages. This continues the trend by introducing
a new abstraction for a locally checked-out abstractions. This will be
used to manage the local cache of packages and to provide a way to
interact with the package contents.