Add `registry.Cache` support in registry
A new `registry.Cache` type was added to the `registry` package. This
leverages the triple of `(registry, package, version)` to cache files
from different registry locations. The system also enables full support
for a listener/observer type, which can be used to monitor and improve
downloads so that the front-end of the CLI can provide better feedback
to the user.