Add support for local registry in the Cache
The `registry.Cache` currently only supports packages that have been
downloaded from a registry somewhere -- but does not enable any form of
override for files that may already be on-disk, but not explicitly
downloaded. This change now introduces an override escape-hatch for this
so that users may specify a path on-disk to a pre-downloaded FHIR
package.

This change deprecates the `fhirsource` abstraction.