Fix download Cucumber test
The `download` cucumber test was broken in the previous change due to
the table of packages being checked including a copy of the `us.core`,
which was never meant to be included.

This removes the duplicate, which (correctly) breaks due to the package
actually being downloaded when it's not meant to be.