Report multi-errors in integration tests
The integration test was previously returning only the _first_ failure
in table-based cases that may contain multiple failures at once.
In order to make it a bit easier to identify the cause of a failure,
this now updates the code to merge all of the failures together into
one succinct output instead.