Improve and fix integration tests
The integration test fixture is failing on `master` for Windows
specifically. The reason, funny enough, is not a Windows-specific
problem -- it's actually that the test-fixture is struggling to
delete the temporarily created directory.

This has been fixed by changing how the `fhenix` binary is installed.
Rather than overriding `GOPATH` with a new temporary path, and
installing all binaries with `go install`, this now does a `go build`
with the binary and uses that instead.

This is done for two reasons:

1. It doesn't force rebuilding the entire project in a new directory,
   which was not previously sharing reused built artifacts
2. It doesn't doesn't require linking to files under the `GOPATH`,
   which was actually the cause of the issue. Deleting the directory
   after it's been linked was resulting in a "file in use" error.

In the process, some new tests have been added.