Add initial draft of `snek` application library
The cobra library is really difficult to use in a nice and encapsulated
way. This saddened and frustrated me to no end; so I designed my own
wrapper around the cobra library to add some sane defaults and make it
easier to work with.

This is still largely untested, but functionally working code. This will
eventually migrate to become its own separate package, but for now lives
as a part of the `fhenix` application.