Add cursor support to snek
For a CLI to look and feel nice for a user, it's important for it to be
able to provide a nice and ergonomic API that is both visually appealing
and easy to understand. To do that, frequently it's desirable to
overwrite lines so that output can appear to "update" in place.

In order to support this, this project has taken on a dependency to the
'cursor' library.