Handle error-printing from commands
By default, Cobra commands print the usage description every time an
error is returned from RunE. This is actually a terrible and largely
nonsensical default, since not all errors need to be accompanied by a
usage message. This commit changes the default behavior to only print
the usage message if the error is of a usage error type.

This required some manual work of disabling the automatic printing of
error messages and calling `Usage` directly.