Return StatusCode rather than exiting internally
The framework currently exits on error manually to prevent Cobra from
printing usage; but this is no longer required with the previous change.
Exiting early actually causes issues with other parts of the language,
like defer calls -- since it means deferrals aren't actually running.

This adjusts it so that a StatusCode is returned instead, which can be
called to `Exit()` directly.