Update application info structure
The application info structure has now learned how to convey both
documentation URLs as well as repository URLs in a more generic manner.
The existing terms have been renamed for the issue URL as well, to now
be `ReportIssueURL`. In doing so, the `panic` template needed to be
updated to reflect the new name.