Fix root JSON schema definition
Copilot did most of the writing for this, and I didn't review
thoroughly enough to catch that the root definition was using a `string`
version field instead of an integer one, as is actually defined in both
v0 and v1.

This wasn't testable until pushed due to it referencing the canonical
URL https://friendly-fhir.org/fhenix/jsonschema/*, which didn't exist
until the last push.

This fixes the issues by using the correct version and type instead,
and addresses an issue in the conditional schema selection.