Allow `Archive` to be uncompressed tarball
The `Archive` type previously only worked with gzipped tarballs, but
this makes it more complex to test since a tarball needs to be gzipped
first, which is multiple layers of wrapping.

To help minimize this, the `Archive` type now supports uncompressed
tarballs as well. In a follow-up commit, support will be added for
testing from a full embedded filesystem instead of requiring checked-in
tarballs.