Rework template funcs
The existing template function approach was ad-hoc on a need-to-have
basis, but this lead to a lot of fragmentation. This change reworks
the whole process to now define a series of top-level "modules" that
provide functionality to the template system.

As part of this, a new mechanism to provide output to errors was also
introduced, which is a "reporter" system for outputting when a
function should have errored. This makes it so that the template is not
responsible for handling the error itself, but it can still be provided
back to the caller as needed.