Add support for obsolete config range
With the recent release of the "v1" config format, the v0 format was
officially made obsolete. It is not unreasonable to expect that this
tool will go through future iterations of schema versions, and there
may be releases that support more than some versions, but not all,
during a migration period.

As a result, this change now adds a minor but simple branch for handling
the official "supported" schema versions, and anything earlier than
that is marked "obsolete".