Add support for terminal line reservation
The existing mechanism for printing to different line in the terminal
was done ad-hoc by just recording the line number and doing the offset
logic manually.

This now adds support for reserving specific terminal lines, which
allows for a more structured approach to printing to different lines.
As a proof-of-concept, the TTY listener was updated to use this new
feature.