Fix download progress bar formatting
The download progress bar was previously pushing the "dots" further
than the line was meant to be printed. The issue was determined to
be due to unicode characters being more than a single byte long -- and
the calculations of the offsets using `len(s)` instead of `utf8.RuneCountInString(s)`.

This updates to use the rune count function instead, and updates the
formatting to be pinned to the length of the terminal.