Trigger usage error for default run command
The default run-command was not triggering any errors for bad input,
which technically allows it to accept any number of arguments without
any error. This commit adds a check for the number of arguments passed
to the default run-command and triggers an error if the number of
arguments is not 0 on the default.