Add package for Set operations
Go mysteriously does not come with a "Set" package, but comes with a
builtin unordered map. This has lead the community to do one of two
different and divergent approaches to managing sets: using a `bool`
as the value, or using a `struct{}` as the value.

This projects takes the decisive step of using a `struct{}` as the
value to represent "no value", which may be different from `bool`
representing _a_ value that may be `true` or `false`.

This has been applied in the `run` function so that template errors
are de-duplicated.