Support multiple package inputs
The current state of the generator is that it only supports a single
package input at a time, which -- overall -- is not a real  issue.
However, it does limit the ability to generate for more than one
package group at a given time.

This change now makes it possible to specify multiple package inputs
in the configuration file, and the generator will generate the
appropriate output for each package input.