Fix --output overriding config file
The --output flag was being ignored whenever the configuration yaml
file defines the `output-dir` config node. Effectively, `--output`
would be fully ignored if the yaml defined an output path -- which was
not intentional.

This now modifies the run command to properly detect this case and
now properly handles overriding the output path. This now means that
`--output` will always take priority, if specified.