Update registrytest package to support `fs.FS`
The fake client and fake server types within `registrytest` currently
work solely on tar.gz file types, which is extremely limiting -- despite
being representative of the FHIR server itself. However, for testing
purposes, it would be way better to not have to check in `tar.gz` files.

This will also play a huge role soon for the integration tests, which
can be modified to work with the fake test server as well. For now, this
makes it possible to read an `embed.FS` as a fake package and it will
operate as if it was a normal tar.gz file.

The existing uses of the gzip model have been updated to primarily
leverage folder-structures instead now.