Ensure integration tests don't run in development
It is extremely easy to trigger 429 rate-limiting errors from Simplifier
thanks to the integration tests in this repo. The VSCode plugin will,
by default, re-trigger to run code-coverage generation regularly --
and this triggers many concurrent reads of the same package.

Generally this should just be a problem with the user's setup; but this
project effectively prescribes using VSCode to simplify things, and it's
an easy issue to trigger as soon as you spawn up the `dev-container`
and look at the tests.

To fix this problem, the `.vscode` settings for the specific _directory_
are being checked in. This does a guaranteed override to ensure that
tests aren't being executed in the background, even if it would otherwise
be reasonable to.

This might be removed in a future commit.