Run explicit command on integration tests
The integration tests were previously running `fhenix` from within the
PATH, which works on everything except Windows for some reason (possibly
due to the `.exe` suffix; it's not clear). As an attempt to fix this
failure and make the tests more robust, this is now replacing fhenix
commands with explicit paths to the binary.

This requires a small translation layer to be done between the setup
of the test-suite and the complete teardown.