Fix bad goreleaser checksums file
The goreleaser configuration that was recently updated is producing
checksum files that end in a newline. This is a valid character on
Linux, but not on most file systems (and arguably, should never be
valid given that it is unreadable). This commit updates the goreleaser
configuration to remove the newline from the checksum template.

In the process, this also adds a few new linker flags to the generated
binaries so that it embeds appropriate version information.