Resolve deadlock scenario
The `release` and `continuous-integration` workflows formed a
"deadlock" scenario according to GitHub, due to both `release` and
`continuous-integration` defining a separate `concurrency` group
that conflicted with one-another.

This commit resolves the deadlock by removing reusing the
`continuous-integration` workflow from the pipeline, and instead only
leveraging the `build-and-test` for long tests.

This is likely better anyway since there is less chance of spurious
failures.