Explicitly specify token for homebrew release
The `goreleaser` tool is an absolutely incredible tool. I mean that
literally, it lacks credibility that this is a real tool.
The documentation is insanely lacking for something that is the de-facto
standard for Go development. It's like the docs were written by a
1st-year college intern.

The majority of the docs for the fields provide you only with cursory
information that you can infer from the configuration node itself, but
it doesn't tell you what most defaults are, or what the behavior is when
set or not set. This is causing a constant struggle to get this damn
thing working.

> But the error message on failure is at least helpful, right?

HAHA, no -- no that would just be too competent. It just tells you 403
with no other information. Is it using the token that everything else
is using? God knows, the docs don't tell you anything.

This desperately tries to cling to what sanity I have left by explicitly
telling it the token to use is `.Env.GITHUB_TOKEN`, and I hope to
whatever deity is listening that this works.

Seriously `goreleaser`, get your crap together. This project is an
open-source project too, but anywhere there is documentation usually
aims to provide GOOD info. Considering `goreleaser` has a paid version,
you should try it sometime.

/rant