Add special handling for x86_64 in install.sh
The install script is currently failing becuase it's written with the
expectation that the arch will be `amd64`. This is not the case, since
we normalize the goreleaser output to use x86_64, which is the more
conventional architecture name.

This adds special handling to allow x86_64 to be used correctly.