Normalize `Windows` -> `windows` in `install.sh`
The `install.sh` script seems to normalize MinGW, Cygwin, etc into
a string of `Windows` for the OS -- but the script later checks that its
in lowercase (`windows`) and fails if it isn't.
This appears to potentially be a bug from the source where this script
was adopted from (`golangci-lint`).

This commit changes the normalization to `windows` to match the
expected casing of the script.