Move public packages to `pkg` directory
In an effort to improve project organization, all of the public packages
which were previously in the root of the project have been moved to a
top-level `pkg` directory. This change is intended to make it easier to
discover the project structure, since there are other important top-level
directories that are package imports -- such as the `jsonschema`
definitions, `docs`, `examples`, etc.

It is undecided if the command will be moved to a corresponding `cmd`
directory as is often done in Go projects.