Fix shell completions for zsh brew installations
The `goreleaser` configuration for the brew tap for this project had
a slight error in the shell completion script. This isn't caught in any
other shell, so it's not a problem -- but will not work properly for
`zsh`, due to the wrong program name being provided.

This commit fixes that issue.