Add new NPM package client
The original design for unpacking logic in `fhenix` was flawed, and lead
to issue #13 where the package percentage is not calculated correctly.
Fixing this requires a literal overhaul of the fetching mechanism.

This aims to be the first step towards that. It doesn't (yet) delete any
code, but it adds a more stable/testable architecture that works with
the existing codebase. This is done by implementing the following:

* A new `npm` package is introduced that contains an `npm.Client`. This
  client is now responsible for fetching packages from the registry,
  and returning unzipped readers to the tar packages.

* In order to intercept the "actual" unpacked size and the current
  progress of the download, a new `ReadInterceptor` function is
  introduced. This function is called when reading bytes from the
  response body, and can be used to compute the percentage of a
  download. This now requires the caller of this client to be built
  with knowledge of this interceptor -- and also means that caching
  and deletions will become the responsibility of an outside
  coordinator.

* A new utility functionality is introduced that makes walking through
  all entries in a reader returned from the `npm.Client` easier.

* A new utility visitor is introduced that can unpack all entries from
  a tarball into a location on disk.

* Finally, to make this all testable, a comprehensive and robust
  test-double package has been introduced: `npmtest`. This package allows
  for configuring "fake" npm registries that are built from `fs.FS`
  implementations -- which may be in-memory, on-disk, etc.

This is a large change, but it is a necessary one to improve the output
of the `fhenix` tool. Long-term, it should also make this much more
testable -- at least in theory.